Move Syncer child modules to their own files
