+++ lib/kennel/syncer.rb
@@ -1,5 +1,10 @@
 # frozen_string_literal: true
 
+require_relative "./syncer/matched_expected"
+require_relative "./syncer/plan_displayer"
+require_relative "./syncer/resolver"
+require_relative "./syncer/types"
+
 module Kennel
   class Syncer
     DELETE_ORDER = ["dashboard", "slo", "monitor", "synthetics/tests"].freeze # dashboards references monitors + slos, slos reference monitors
@@ -161,222 +166,6 @@ module Kennel
           tracking_id = a.fetch(:tracking_id)
           !tracking_id || tracking_id.start_with?(*project_prefixes)
         end
-      end
-    end
-
-    class PlanDisplayer
-      def initialize
-        @attribute_differ = AttributeDiffer.new
-      end
-
-      def display(internal_plan)
-        Kennel.out.puts "Plan:"
-        if internal_plan.empty?
-          Kennel.out.puts Console.color(:green, "Nothing to do")
-        else
-          print_changes "Create", internal_plan.creates, :green
-          print_changes "Update", internal_plan.updates, :yellow
-          print_changes "Delete", internal_plan.deletes, :red
-        end
-      end
-
-      private
-
-      def print_changes(step, list, color)
-        return if list.empty?
-        list.each do |item|
-          Kennel.out.puts Console.color(color, "#{step} #{item.api_resource} #{item.tracking_id}")
-          if item.class::TYPE == :update
-            item.diff.each { |args| Kennel.out.puts @attribute_differ.format(*args) } # only for update
-          end
-        end
-      end
-    end
-
-    class Resolver
-      def initialize(expected:, project_filter:, tracking_id_filter:)
-        @id_map = IdMap.new
-        @project_filter = project_filter
-        @tracking_id_filter = tracking_id_filter
-
-        # mark everything as new
-        expected.each do |e|
-          id_map.set(e.class.api_resource, e.tracking_id, IdMap::NEW)
-          if e.class.api_resource == "synthetics/tests"
-            id_map.set(Kennel::Models::Monitor.api_resource, e.tracking_id, IdMap::NEW)
-          end
-        end
-      end
-
-      def add_actual(actual)
-        # override resources that exist with their id
-        project_prefixes = project_filter&.map { |p| "#{p}:" }
-
-        actual.each do |a|
-          # ignore when not managed by kennel
-          next unless tracking_id = a.fetch(:tracking_id)
-
-          # ignore when deleted from the codebase
-          # (when running with filters we cannot see the other resources in the codebase)
-          api_resource = a.fetch(:klass).api_resource
-          next if
-            !id_map.get(api_resource, tracking_id) &&
-            (!project_prefixes || tracking_id.start_with?(*project_prefixes)) &&
-            (!tracking_id_filter || tracking_id_filter.include?(tracking_id))
-
-          id_map.set(api_resource, tracking_id, a.fetch(:id))
-          if a.fetch(:klass).api_resource == "synthetics/tests"
-            id_map.set(Kennel::Models::Monitor.api_resource, tracking_id, a.fetch(:monitor_id))
-          end
-        end
-      end
-
-      def resolve_as_much_as_possible(expected)
-        expected.each do |e|
-          e.resolve_linked_tracking_ids!(id_map, force: false)
-        end
-      end
-
-      # loop over items until everything is resolved or crash when we get stuck
-      # this solves cases like composite monitors depending on each other or monitor->monitor slo->slo monitor chains
-      def each_resolved(list)
-        list = list.dup
-        loop do
-          return if list.empty?
-          list.reject! do |item|
-            if resolved?(item.expected)
-              yield item
-              true
-            else
-              false
-            end
-          end ||
-            assert_resolved(list[0].expected) # resolve something or show a circular dependency error
-        end
-      end
-
-      private
-
-      attr_reader :id_map, :project_filter, :tracking_id_filter
-
-      # TODO: optimize by storing an instance variable if already resolved
-      def resolved?(e)
-        assert_resolved e
-        true
-      rescue UnresolvableIdError
-        false
-      end
-
-      # raises UnresolvableIdError when not resolved
-      def assert_resolved(e)
-        e.resolve_linked_tracking_ids!(id_map, force: true)
-      end
-    end
-
-    module MatchedExpected
-      class << self
-        def partition(expected, actual)
-          lookup_map = matching_expected_lookup_map(expected)
-          unmatched_expected = Set.new(expected) # for efficient deletion
-          unmatched_actual = []
-          matched = []
-          actual.each do |a|
-            e = matching_expected(a, lookup_map)
-            if e && unmatched_expected.delete?(e)
-              matched << [e, a]
-            else
-              unmatched_actual << a
-            end
-          end.compact
-          [matched, unmatched_expected.to_a, unmatched_actual]
-        end
-
-        private
-
-        # index list by all the thing we look up by: tracking id and actual id
-        def matching_expected_lookup_map(expected)
-          expected.each_with_object({}) do |e, all|
-            keys = [e.tracking_id]
-            keys << "#{e.class.api_resource}:#{e.id}" if e.id
-            keys.compact.each do |key|
-              raise "Lookup #{key} is duplicated" if all[key]
-              all[key] = e
-            end
-          end
-        end
-
-        def matching_expected(a, map)
-          klass = a.fetch(:klass)
-          map["#{klass.api_resource}:#{a.fetch(:id)}"] || map[a.fetch(:tracking_id)]
-        end
-      end
-    end
-
-    module Types
-      class PlannedChange
-        def initialize(klass, tracking_id)
-          @klass = klass
-          @tracking_id = tracking_id
-        end
-
-        def api_resource
-          klass.api_resource
-        end
-
-        def url(id = nil)
-          klass.url(id || self.id)
-        end
-
-        def change(id = nil)
-          Change.new(self.class::TYPE, api_resource, tracking_id, id)
-        end
-
-        attr_reader :klass, :tracking_id
-      end
-
-      class PlannedCreate < PlannedChange
-        TYPE = :create
-
-        def initialize(expected)
-          super(expected.class, expected.tracking_id)
-          @expected = expected
-        end
-
-        attr_reader :expected
-      end
-
-      class PlannedUpdate < PlannedChange
-        TYPE = :update
-
-        def initialize(expected, actual, diff)
-          super(expected.class, expected.tracking_id)
-          @expected = expected
-          @actual = actual
-          @diff = diff
-          @id = actual.fetch(:id)
-        end
-
-        def change
-          super(id)
-        end
-
-        attr_reader :expected, :actual, :diff, :id
-      end
-
-      class PlannedDelete < PlannedChange
-        TYPE = :delete
-
-        def initialize(actual)
-          super(actual.fetch(:klass), actual.fetch(:tracking_id))
-          @actual = actual
-          @id = actual.fetch(:id)
-        end
-
-        def change
-          super(id)
-        end
-
-        attr_reader :actual, :id
       end
     end
   end
