+++ test/kennel/syncer/resolver_test.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative "../../test_helper"
+
+# Covered by syncer_test.rb
+SingleCov.covered! uncovered: 31
